Keep shared runner UI after model change
Even though we are now counting minutes when shared runners are
disabled, we don't want to show that information in the UI, since it may
be confusing to users to see minutes being counted after they have
disabled shared runners. Avoiding changes to the UI involves re-adding a
check for Namespace#any_project_with_shared_runners_enabled? in some
places.