Move the container registry visibility setting
Currently, the Container Registry visibility setting is nested within
the Repository visibility setting. But on the backend, Container
Registry visibility is not affected by Repository visibility.

This commit moves the Container Registry visibility setting so that
it is not nested within the Repository setting, and the input element
is not affected by the Repository setting.