Report to Sentry for long Redis durations
We're seeing some very high (over 5 seconds) Redis durations in our
logs, and we don't know what's causing them. This is an attempt to
answer that by sending an event to Sentry when we detect this
happening.

It's behind the feature flag `report_on_long_redis_durations`, which is
off by default.

starting