Allow prevent_cross_joins to be used directly in tests
This method can be passed a block and we can validate that no SQL query
in that code block attempts to join across ci/main databases. This is
useful in the intermediate time while we attempt to fix many small
issues until we enable this as the default in all RSpec tests.