Update CI Minutes eligibility check to use new API
We want to use the new API for checking if a namespace is eligible for
CI minutes as the previous one assumed that the purchase check was for
new license subscriptions. This would filter out namespaces that have a
Premium/Ultimate subscription, when those namespaces are actually
eligible for a CI minutes purchase.

Update the subscriptions controller to fetch and supply the plan ID to
this API.

Issue: https://gitlab.com/gitlab-org/gitlab/-/issues/326663
MR: https://gitlab.com/gitlab-org/gitlab/-/merge_requests/66981