Remove `ref_name` parameter in `RemoteMirror`
On-disk remotes are deprecated, and in fact we don't create any remotes
in GitLab anymore in favor of using in-memory remotes, where we just
pass the URL to Gitaly. We thus don't need to pass information about
remote names anymore given that it's unused nowadays.

Remove the `ref_name` parameter in `RemoteMirror` to reflect this. Note
that this parameter is misnamed, and should really have been called
`remote_name`. Furthermore, drop now-unneeded infrastructure which was
required to track the remote name.