Add `container_policy` to policies controller
For MVC, we would like for the policies controller to be able to
return both `scan_execution_policy` and `container_policy`
types. Until the container runtime policies are migrated to use
a security policy project, we will call out to the
NetworkPolicies::FindResourceService as a workaround, using some of the
same logic as the ThreatMonitoringController. The
ThreatMonitoringController will be deleted when all of the functionality
in this controller is completed.