Fix specs that use invalid arguments to Gitlab::Git::Commit.between
Unlike the CommitsBetween RPC, the ListCommits RPC will issue an error
if an unknown or invalid revision is given as an argument. This should
generally be fine in production, where (e.g.) attempting to view
changes on a branch that has just been deleted can justifiably error
out.

In the test suite, we use stub revisions a lot, and where these get to
Gitlab::Git::Commit.between, they cause an error in previously-passing
specs. We don't rely on the result of `between` in any of these specs,
so we just need to fix the tests not to trigger the behaviour.