Use v-safe-html for field help texts in integration form
This isn't critical since the help texts are defined in the integration
models (in `Integration#help` and `Integration#fields`) and can be
trusted.

We still want to allow external links using `target="blank"`, so we pass
a custom configuration to `v-safe-html` until this is implemented in
`gitlab-ui` with https://gitlab.com/gitlab-org/gitlab-ui/-/issues/1427.

We also do the same for the main help texts in `integration_form.vue`,
and add the `target` and `rel` attributes where missing.