Add retry mechanism for proxied Jira installations
This adds a queue for proxied JiraConnect installation event requests.
Failures cause new attempts to be enqueued, up to three times. This
ensures that we can retry, without tying up a worker thread, and
allowing the attempts to be spread out over time, since if a remote
server is experiencing downtime, we would anticipate immediate retries
would be less likely to succeed.