Allow to create epic from ancestor board
When an epic is created from an epic board, user can
can select a group where the epic will be created. This group
can be either the same group as board's group or any subgroup.

For this reason when we are searching for the board, it can exist in
any ancestor group.

Changelog: added
EE: true