Use a SSOT to check if service ping is enabled
Instead of checking the application settings directly, this is abstracted to
a single source of truth.

For projects with licenses, in EE we check if the license requires tracking.

With this change, events will be stored locally even if the service ping is
disabled for licensed users. These events will not be sent as part of the service
ping if the service ping is disabled.