Remove broken "default" storage initialiser
This initialiser was trying to setup the "default" storage if it didn't
exist in the config file but since the settingslogic tries to validate
that back to gitlab.yml it finds the setting doesn't exist and blows up.

Since this code doesn't actually work and since this is one part of the
reason why users cannot run gitlab without a "default" storage we should
remove it.

Changelog: fixed