Add ResetJobTokenScopeEnabledAgain migration
This migration drops the `job_token_scope_enabled` field on
project_ci_cd_settings` and re-adds it. This will effectively cause its
value for all projects to be `false`, which will let us safely re-enable
the `ci_scoped_job_token` feature flag.

We previously reset this field before turning on the feature flag for
the first time. We are now resetting it again because we had to turn off
the feature flag after an incident.

Changelog: other