Always enable the database load balancer
By always enabling the database load balancer we reduce the amount of
different code paths, make it easier for developers to interact with the
load balancing code (e.g. by adding metrics), and make it easier to add
support for multiple databases (e.g. by not having to check if the load
balancer is enabled or not).

As part of this commit we fixed various parts of the code that didn't
expect the load balancer to always be enabled (or at least when running
tests).

Changelog: added
Co-Authored-By: Dylan Griffith <dyl.griffith@gmail.com>