Always use the newest commit in a branch for push events
Prior to this commit, when creating a new default branch, the push
event would reference the *oldest* commit in the branch, rather than
the newest. This is the opposite of what we do when creating any other
branch, or when updating any branch, whether it's the default branch or
not.

Even worse, when the new default branch contained more than 100 commits
we would return the 100th commit, rather than the oldest. This is
unambiguously a bug - that behaviour is not useful at all.

Fix the ordering of commits when pulling data from the default branch
so we always use the newest one when creating the default branch.

Changelog: fixed