Clarify the milestone requirements for security MRs
Security releases happen after the main monthly release, so the first
"backport" will actually be for the most recent release.

This frequently causes confusion and authors end up creating an
unnecessary backport for an older release, instead of the latest one.

This is explained in the [security workflow documentation][1], but we
should also clarify this in the security issue and merge request
templates.

We don't really need a milestone on the MR merging into `master`, so
we also change the Danger rules to not require one.

[1]: https://gitlab.com/gitlab-org/release/docs/blob/master/general/security/developer.md#security-merge-requests