Make notification_email like commit_email
When there is no database value for the User#commit_email attribute,
the User model returns the primary email instead. This modifies the
User#notification_email attribute to behave the same way, instead of
setting the database value as if there had been user input. The intent
of this change is to help standardize the behavior of User secondary
email attributes, as part of a broader refactoring.