Geo: Replicate wiki and design repository HEAD ref
If a design or wiki repository on a secondary site has a HEAD ref that
does not match the HEAD ref on the primary site, then resync will make
it match.

If verification fails due to a mismatching HEAD ref, then a resync will
allow verification to succeed.

Changelog: fixed
EE: true