Remove an ineffective wait-loop in a QA spec
This code was added to try to prevent a "branch not found" error when
processing push options. However, this is literally a race condition in
the application that needs to be fixed, rather than something to work
around.

Additionally, the git clone operation used bypasses all the caching in
the Rails application, so is not a useful check. If we wanted to keep
this check, we could loop around `QA::Resource::Project.has_branch?`
instead, but it's much better to do without it.