Refactor sidebar menus link method
Most of the sidebar menus default to the first menu item link as the
main menu link.

Instead of writing it in every menu, we can just move it to the
base menu class and let the other menus inherit it. We'd just
need to override those with a different behavior.