Fix bug validating EE project features
The current class `ProjectFeature` validates that none of the
project features have a value greater than `Featurable::ENABLED`,
except for the `pages` feature.

Nevertheless, we have an EE extension where we add more features to
the `ProjectFeature` class. The current validator use only the
constant `ProjectFeature::FEATURES` for the validation. This means
that the `ProjectFeature::EE_FEATURES` are not included in the
validation and we could set the `Featurable::PUBLIC` value.

In this commit we fix that by including all features set in the
`ProjectFeature` including both FOSS and EE ones.

Changelog: fixed