Fix SSO SAML redirection not including query string
When a user without a SSO session attempted to access anything in a SAML
group with a query string, previously GitLab would redirect the user
back to original path but drop the query string.

We fixed the redirection originally in
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/66791, but
`request.path_info` drops query strings. To ensure the query strings are
passed to the `RelayState`, we need to use `request.fullpath`.

Relates to https://gitlab.com/gitlab-org/gitlab/-/issues/338833

Changelog: fixed