Reduce Gitaly calls for keeping around refs of published notes
We are currently keeping around commits of each note being
published even though they are all the same commits. That results
to duplicate `ref_exists` calls.

For example, if there are 6 draft notes set to be published on a
single MR version, there will be around 36 calls for `ref_exists`.

This fix reduces those calls by getting the unique shas to be kept.
Based on local testing, this reduced the `ref_exists` calls from
36 to 4.

Changelog: performance