Use dynamic mapping in trigger to sync `integrations.type_new`
In https://gitlab.com/gitlab-org/gitlab/-/merge_requests/66541 we ended
up going with a static mapping for this trigger, and figured any new
integrations could just start using the correct class name in `type`
directly.

But our plan is to drop the `type` column at some point and replace it
with `type_new`, so we still need to sync the class name to `type_new`
for new integrations as well.

So the trigger is changed to transform all `*Service` values into
`Integrations::*`, and copy any other values unmodified.

Changelog: fixed