merge_request: Drop checks whether a squash is in progress
Originally, Gitaly has implemented squashes via a separate worktree of
the repository. This has changed though in v14.1, where Gitaly now uses
an in-memory implementation of squashes. The result of this is that it
is both faster (~4x) and less error-prone.

One side effect though is that it's not possible to tell anymore whether
a squash is in progress or not: there is no on-disk state anymore which
Gitaly could check. Because of this, Gitaly has deprecated the RPC and
will eventually remove it.

Remove this check on GitLab's side to adapt to this change: it doesn't
work anymore anyway. If it needs to resurface, the check would instead
need to be implemented in GitLab itself, where any ongoing squashes
would need to be recorded as part of the database.

Changelog: removed