Remove approve in External Status Checks API docs
To avoid confusion with Merge Request approvals, and to maintain a
distinction between the value of human and bot judgments, we do not
use the term "approve" or "approval" in relation to External Status
Checks. This updates the API documentation to remove the term from
a few locations, because it was briefly in use.