Disable IMAP auth check in Rake task when Microsoft Graph is configured
MailRoom supports two inbox modes: IMAP and Microsoft Graph. The former
uses the IMAP protocol, and the latter uses a REST API to retrieve
e-mail.

When we added support for Microsoft Graph, we did not account for this
in the `gitlab:incoming_email:check` Rake task.

For now, disable the IMAP check if is it not configured. We may want to
consider adding a Microsoft Graph-specific check, but MailRoom will exit
if the Microsoft Graph credentials are not configured properly.

Relates to
https://gitlab.com/gitlab-org/gitlab/-/issues/214900#note_656345704