Group Settings CI/CD h4 expand
Group Settings / CI/CD headers General pipelines, Runners, and Auto
DevOps (h4 elements) do not expand/collapse on mouse-click / tap, but
only using the Expand/Collapse button. This may seem inconsistent to
the user.

This MR adds classes settings-title, js-settings-toggle and
js-settings-toggle-trigger-only to the h4 headers in Group Settings /
CI/CD. This enables the expand/collapse behavior on on-click/on-tap for
the h4 headers.

Changelog: changed