Update relative positions on querying board issues
When board issues are queried via graphql,
check and initialize the relative positions of issues
if they are null.

Why?

The issues appearing in a board list must not have
their relative positions as `nil` to allow
repositioning of the issues.

Boards::IssuesController set the relative positions
of the issues being fetched but we did not do the same
when board issues were queried via graphql.

Changelog: fixed