Fix Epic bulk updates leaking to other epics
When doing a bulk update on epics, the set of labels for the first
selected epic are copied over to the rest of the epics.

This is because we compute the final list of labels and sets it to
`params` and this `params` is the same one used for updating other
epics.

Changelog: fixed
EE: true