checks: Convert PushRuleChecks to allow for batching
The PushRuleChecks are implemented as BaseSingleChecker, where they
always operate on a single change only. This is inefficient for a subset
of push rules, most namely for the file size rules: we could load all
blobs via a single RPC call and inspect their sizes instead of doing an
RPC call per change.

Prepare for this change by converting the PushRules to be implemented as
BaseBulkChecker. For now, nothing really changes: we just deconstruct
the bulk changes into SingleChangeAccesses and then pass those into our
subrules. But with this change in place, we can easily convert subsets
of our push rules to use batching.