checks: Make push rules available in BaseBulkChecker
Push rules are only available in the Enterprise Edition and are thus
mixed into the BaseSingleChecker via a separate EE module which extends
the base class. This only makes push rules available for this single
checks: bulk checks ain't got any access to them, and until now they
haven't needed any either. We're about to refactor push rules to allow
batching of access checks though, so we need access to push rules in
both classes.

Move the push rule implementation into the BaseChecker class. This is
the parent class of both BaseSingleChecker and BaseBulkChecker, so it
will make the function available to both.