Add param to allow scoped caching of Repo#merge_to_ref
Adding caching to this endpoint for all use cases triggered a bug in an
untested portion of Merge Trains. Adding this param will allow us to
selectively apply caching, improving the performance characteristics of
where we have a problem (MergeRequests::MergeabilityCheckService) but
not in the problematic Merge Trains code.

Changelog: added