Prepopulate new issue with link to the parent
At the top of each issue detail view, there's a "New issue" link in the
dropdown menu.  Modify that link to prepopulate the newly-created issue
with a link to the original issue, to make it easier to create
interconnected issues without having to copy-paste or memorize the
parent issue id.

Changelog: changed