Add default option to notification_email input
The User#notification_email attribute is assumed to be the primary
email when not set. Historically the notification preferences form has
automatically set an input value of the current primary email. However,
recent refactoring has made this problematic. The notification_email is
no longer proactively set in the database to match the primary email.
Instead, we now make default assumptions in the code, and want to rely
on user input for database values. Toward that goal, this introduces a
new default option to the notification_email input. It is a blank
option, meaning the default assumption in the code will be used (the
primary email). Now the notification_email value will not be set in
the database unless the user provides input requesting that.

Changelog: fixed