Remove scanner_type argument from GraphQL mutation
There's no such a thing as scanner_type anywhere in the database. We
have a report_type which we're setting to `generic` if the Vulnerability
is created via this GraphQL migration. This was agreed on in
https://gitlab.com/gitlab-org/gitlab/-/issues/10272#note_577520606

This doesn't need to follow any deprecation process since it wasn't
released yet and is under a feature flag.

Changelog: changed
EE: true