checks: Add function to convert ChangesAccess to SingleChangeAccesses
There is only a single callsite right now which extracts changes from a
ChangesAccess and converts them into SingleChangeAccesses, which is
inside the ChangesAccess class itself. We're about to convert push rules
such that they can both drive single and bulk push rules, and there
we'll need the same functionality.

Extract the code into a standalone `#single_change_accesses()` function
which can be used to do the conversion.