Only set User#commit_email with user input
Previously the User profile form automatically set a value for
User#commit_email. This would become the database value when the
user submitted the form, whether the user was aware of this or not.
That behavior was unnecessary, because the User model already
assumes the commit_email is the primary email when not set. Further,
this behavior complicated validation when the primary email is
being changed. As a simple solution, this change introduces a blank
option to the select dropdown, which is used by default. Now the
database value will only change if the user changes it.

Changelog: fixed