Create new setting for user deactivation emails
With the introduction of:
https://gitlab.com/gitlab-org/gitlab/-/merge_requests/64964

Email notifications are sent to users when their account is deactivated.
This was an automatic change to behavior without an option to disable.

Administrators often deactivate users to control instance license usage.
Sending an email to users can defeat the purpose of deactivation since
an alerted user can reactivate their account and consume a license.

A new setting enables administrators to control whether or not email
notifications will be sent to users upon account deactivation.

Changelog: added

Signed-off-by: Vincent Fazio <vfazio@xes-inc.com>