Temporarily rescue Module::DelegationError in worker
Groups::ExportMembershipsWorker is currently raising
Module::DelegationError in circumstances where User records are
unavailable. That is causing these workers to fail and be re-tried,
only to fail again. To avoid this impact on infrastructure while we
troubleshoot the bug, this adds a temporary rescue clause to the
worker. This clause will be removed when the bug is fixed in a
follow-up.