Patch `grape-entity` to prevent having NameError loop
RubyVM starts raising unlimited amount of `NameError` errors in case of
a `NameError` in our Grape::Entity instances because of a bug exists in
the `grape-entity` gem.

This causes high CPU usage and puts us at the risk of having DoS.
The VM would eventually raise a `SystemStackError` but this takes quite
long time because the repeating frames are huge.

You can track the progress of upstream fix here:
https://github.com/ruby-grape/grape-entity/pull/355

Changelog: fixed